requirements.txt
# Hermes AGI
fire
pyyaml
requests
rich
prompt_toolkit
pydantic>=2.0
openai
python-dotenv
PyJWT[crypto]
httpx
jinja2
tenacity
